Move edit commands to a new file
